fixes + tests

+++ ft_strncmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ebayat <ebayat@student.42.fr>              +#+  +:+       +#+        */
+/*   By: elie <elie@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 13:53:52 by ebayat            #+#    #+#             */
-/*   Updated: 2023/10/05 14:33:46 by ebayat           ###   ########.fr       */
+/*   Updated: 2023/11/05 12:55:45 by elie             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	int	i;
 
 	i = 0;
-	while (n > 0 && s1[i] && s2[i] && s1[i] == s2[i] && n > i)
+	while (n > 0 && s1[i] && s2[i] && s1[i] == s2[i] && (int)n > i)
 		i++;
 	return (s1[i] - s2[i]);
 }
